error

+++ manage-material/src/main/java/vn/com/devmaster/service/managematerial/reponsitory/CartItemRepository.java
@@ -3,5 +3,5 @@ package vn.com.devmaster.service.managematerial.reponsitory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import vn.com.devmaster.service.managematerial.dommain.CartItem;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
 }
